Finds a path to a goal state
Contains different search algorithms that expand the starting state to
its sub-states. These sub-states are expanded and this process
continues until there is a state that is identified as a goal state.
when the state is reached, the list of actions that were taken to reach
that state is returned.